+++ packages/rx_bloc_cli/mason_templates/bricks/feature_widget_toolkit/__brick__/lib/l10n/network_lookup/widget_toolkit_lookup.dart
@@ -1,0 +1,36 @@
+{{> licence.dart}}
+
+import '../../assets.dart';
+import '../{{project_name}}_app_i18n.dart';
+import 'util.dart';
+
+class AppI18nWidgetToolkitLookup extends I18nFeatureWidgetToolkitLookup {
+  @override
+  String getString(String key, [Map<String, String>? placeholders]) {
+    return getFromNetwork('widget_toolkit__', key, placeholders) ??
+        getFromBundle(AppI18n.locale?.languageCode)
+            .getString(key, placeholders)!;
+  }
+
+  I18nFeatureWidgetToolkit? bundledI18n;
+
+  I18nFeatureWidgetToolkit getFromBundle(String? languageCode) {
+    if (bundledI18n == null) {
+      switch (I18n.currentLocale?.languageCode) {
+        case 'bg':
+          bundledI18n =
+              I18nFeatureWidgetToolkit(I18nFeatureWidgetToolkitLookup_bg());
+          break;
+        case 'en':
+          bundledI18n =
+              I18nFeatureWidgetToolkit(I18nFeatureWidgetToolkitLookup_en());
+          break;
+        default:
+          bundledI18n =
+              I18nFeatureWidgetToolkit(I18nFeatureWidgetToolkitLookup_en());
+          break;
+      }
+    }
+    return bundledI18n!;
+  }
+}