+++ packages/rx_bloc_cli/mason_templates/bricks/rx_bloc_base/__brick__/bin/server/utils/utilities.dart
@@ -4,6 +4,8 @@ import 'dart:math';
 
 import 'package:shelf/shelf.dart';
 
+import 'api_controller.dart';
+import 'dependency_injector.dart';
 import 'response_builder.dart';
 import 'server_exceptions.dart';
 
@@ -59,3 +61,24 @@ Handler buildSafeHandler(Function callback, ResponseBuilder responseBuilder) =>
 String generateRandomString([int charsNum = 64]) =>
     String.fromCharCodes(Iterable.generate(
         charsNum, (_) => _chars.codeUnitAt(_random.nextInt(_chars.length))));
+
+/// Helper method for easily configuring dependencies and routes
+Future<({RouteGenerator routeGenerator, DependencyInjector di})>
+    configureRoutesAndDependencies(
+  Future<void> Function(RouteGenerator routeGenerator, DependencyInjector)
+      controllerBuilder,
+  Future<void> Function(DependencyInjector)? dependencyBuilder,
+) async {
+  // Generate dependencies (if any)
+  final di = DependencyInjector();
+  if (dependencyBuilder != null) {
+    await dependencyBuilder.call(di);
+  }
+
+  // Now generate routes
+  final rg = RouteGenerator();
+  await controllerBuilder(rg, di);
+
+  // Return record with given generated values
+  return (routeGenerator: rg, di: di);
+}