+++ packages/rx_bloc_cli/mason_templates/bricks/rx_bloc_base/__brick__/lib/lib_translations/models/i18n_models.g.dart
@@ -1,0 +1,18 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+part of 'i18n_models.dart';
+
+// **************************************************************************
+// JsonSerializableGenerator
+// **************************************************************************
+
+I18nSections _$I18nSectionsFromJson(Map<String, dynamic> json) => I18nSections(
+      item: (json['translations'] as Map<String, dynamic>).map(
+        (k, e) => MapEntry(k, Map<String, String>.from(e as Map)),
+      ),
+    );
+
+Map<String, dynamic> _$I18nSectionsToJson(I18nSections instance) =>
+    <String, dynamic>{
+      'translations': instance.item,
+    };